Change database and services dependency
