+++ backend/database.js
@@ -2,7 +2,8 @@ require("dotenv").config();
 const { Pool } = require("pg");
 
 // Documentation for pg: https://node-postgres.com/
-//load environment variables
+
+// Load environment variables
 const config = {
   user: process.env.DATABASE_USER,
   host: process.env.DATABASE_HOST,
@@ -13,94 +14,22 @@ const config = {
 
 const pool = new Pool(config);
 
-pool.query("SELECT NOW()", (err, res) => {
-  if (err) {
-    console.error(err);
-  } else {
-    console.log("Connection successful");
-  }
-  //pool.end();
-});
-
-////////////////////////////////////////////////////////////////
-
-/**
- * QUERIES THAT DEAL WITH USERS
- */
-
-//get users
-const getUsers = async (request, response) => {
+const transact = async (callback) => {
+  const client = await pool.connect();
   try {
-    const result = await pool.query("SELECT * FROM users ORDER BY id ASC");
-    response.status(200).json(result.rows);
+    await client.query('BEGIN');
+    const result = await callback(client.query.bind(client));
+    await client.query('COMMIT');
+    return result;
   } catch (error) {
+    await client.query('ROLLBACK');
     throw error;
+  } finally {
+    client.release();
   }
 };
-// get a single user
-const getUserById = (request, response) => {
-  const id = parseInt(request.params.id);
 
-  pool.query("SELECT * FROM users WHERE id = $1", [id], (error, result) => {
-    if (error) {
-      throw error;
-    }
-    response.status(200).json(result.rows);
-  });
-};
-
-// create single user
-const createUser = (request, response) => {
-  // should we store password?
-  const { email, password } = request.body;
-
-  pool.query(
-    "INSERT INTO users (email , password) VALUES ($1,$2)",
-    [email, password],
-    (error, result) => {
-      if (error) {
-        throw error;
-      }
-      response.status(201).send(`User added with ID: ${result.insertId}`);
-    }
-  );
-};
-
-// update single user
-// /users/:id
-const updateUser = (request, response) => {
-  // should we store password?
-  const { id } = request.params.id;
-  const { email, password } = request.body;
-
-  pool.query(
-    "UPDATE users SET email = $1, password = $2",
-    [email, password],
-    (error, result) => {
-      if (error) {
-        throw error;
-      }
-      response.send(200).send(`user modified with id: ${id}`);
-    }
-  );
-};
-
-const deleteUser = (request, response) => {
-  const id = parseInt(request.params.id);
-
-  pool.query("DELETE FROM users WHERE id = $1", [id], (error, result) => {
-    if (error) {
-      throw error;
-    }
-    response.status(200).send(`User deleted with ID: ${id}`);
-  });
-};
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 module.exports = {
-  getUsers,
-  getUserById,
-  createUser,
-  updateUser,
-  deleteUser
-};
+  query: pool.query.bind(pool),
+  transact, 
+}
