--- backend/db.js
@@ -1,35 +1,0 @@
-require("dotenv").config();
-const { Pool } = require("pg");
-
-// Documentation for pg: https://node-postgres.com/
-
-// Load environment variables
-const config = {
-  user: process.env.DATABASE_USER,
-  host: process.env.DATABASE_HOST,
-  database: process.env.DATABASE_NAME,
-  password: process.env.DATABASE_PASSWORD,
-  port: process.env.DATABASE_PORT
-};
-
-const pool = new Pool(config);
-
-const transact = async (callback) => {
-  const client = await pool.connect();
-  try {
-    await client.query('BEGIN');
-    const result = await callback(client.query.bind(client));
-    await client.query('COMMIT');
-    return result;
-  } catch (error) {
-    await client.query('ROLLBACK');
-    throw error;
-  } finally {
-    client.release();
-  }
-};
-
-module.exports = {
-  query: pool.query.bind(pool),
-  transact, 
-}
