+++ backend/services.js
@@ -1,0 +1,102 @@
+const { query, transact } = require('./database');
+
+query("SELECT NOW()", (err, res) => {
+  if (err) {
+    console.error(err);
+  } else {
+    console.log("Connection successful");
+  }
+});
+
+////////////////////////////////////////////////////////////////
+
+/**
+ * QUERIES THAT DEAL WITH USERS
+ */
+
+//get users
+const getUsers = async (request, response) => {
+  try {
+    const result = await query("SELECT * FROM users ORDER BY id ASC");
+    response.status(200).json(result.rows);
+  } catch (error) {
+    throw error;
+  }
+};
+// get a single user
+const getUserById = (request, response) => {
+  const id = parseInt(request.params.id);
+
+  query("SELECT * FROM users WHERE id = $1", [id], (error, result) => {
+    if (error) {
+      throw error;
+    }
+    response.status(200).json(result.rows);
+  });
+};
+
+// create single user
+const createUser = (request, response) => {
+  // should we store password?
+  const { email, password } = request.body;
+
+  query(
+    "INSERT INTO users (email , password) VALUES ($1,$2)",
+    [email, password],
+    (error, result) => {
+      if (error) {
+        throw error;
+      }
+      response.status(201).send(`User added with ID: ${result.insertId}`);
+    }
+  );
+};
+
+// update single user
+// /users/:id
+const updateUser = (request, response) => {
+  // should we store password?
+  const { id } = request.params.id;
+  const { email, password } = request.body;
+
+  query(
+    "UPDATE users SET email = $1, password = $2",
+    [email, password],
+    (error, result) => {
+      if (error) {
+        throw error;
+      }
+      response.send(200).send(`user modified with id: ${id}`);
+    }
+  );
+};
+
+const deleteUser = (request, response) => {
+  const id = parseInt(request.params.id);
+
+  query("DELETE FROM users WHERE id = $1", [id], (error, result) => {
+    if (error) {
+      throw error;
+    }
+    response.status(200).send(`User deleted with ID: ${id}`);
+  });
+};
+
+const exampleTransaction = async () => {
+  const result = await transact(async (query) => {
+    const time = (await query('SELECT NOW()')).rows[0];
+    console.log(`Time: ${JSON.stringify(time)}`);
+    const t = await query('SELECT 1');
+    return t;
+  });
+  console.log(result.rows[0]);
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+module.exports = {
+  getUsers,
+  getUserById,
+  createUser,
+  updateUser,
+  deleteUser
+};
